Move all inferrer-related functions and classes to corresponding file
