+++ src/python/spectrseqtools/common.py
@@ -2,9 +2,9 @@ from pathlib import Path
 from typing import List, Tuple
 
 from spectrseqtools.prediction.composition_inference import (
+    CompositionInferrer,
     infer_compositions_with_matrix,
 )
-from spectrseqtools.prediction.traceback_matrix import CompositionInferrer
 
 ERROR_METHOD = "l1_norm"
