+++ src/python/spectrseqtools/prediction/prediction.py
@@ -9,10 +9,12 @@ from spectrseqtools.common import (
     calculate_error_threshold,
 )
 from spectrseqtools.dataclasses import Prediction, SolverParameters
-from spectrseqtools.prediction.composition_inference import is_valid_mass
+from spectrseqtools.prediction.composition_inference import (
+    CompositionInferrer,
+    is_valid_mass,
+)
 from spectrseqtools.prediction.sequence_inference import LinearProgramInstance
 from spectrseqtools.prediction.skeleton_building import SkeletonBuilder
-from spectrseqtools.prediction.traceback_matrix import CompositionInferrer
 
 
 class Predictor:
